Module Cmdliner

Declarative definition of command line interfaces.

Consult the tutorial, the cookbook, program blueprints and source structure, details about the supported command line syntax and examples of use.

Open the module to use it, it defines only these modules in your scope.

module Manpage : sig ... end

Man pages.

module Term : sig ... end

Terms.

module Cmd : sig ... end

Commands.

module Arg : sig ... end

Terms for command line arguments.