Module Cmd.Env

Environment variable and their information.

Environment variables

type var = string

The type for environment variable names.

Environment variable information

type info

The type for environment variable information.

val info : ?deprecated:string -> ?docs:Manpage.section_name -> ?doc:string -> var -> info

info ~docs ~doc var describes an environment variable var such that:

doc is the man page information of the environment variable, defaults to "See option $(opt).".

docs is the title of the man page section in which the environment variable will be listed, it defaults to Cmdliner.Manpage.s_environment.

deprecated, if specified the environment variable is deprecated. Use of the variable warns on depstderr This message which should be a capitalized sentence is preprended to doc and output on standard error when the environment variable ends up being used.

In doc and deprecated the documentation markup language can be used with following variables:

$(opt), if any the option name of the argument the variable is looked up for.

$(env), the value of var.

The variables mentioned in the doc string of Cmd.info.

val info_var : info -> var

info_var info is the variable described by info.