Module Cmdliner_def.Arg_info

Argument information.

type absence =

| Err
(*
an error is reported.
*)

| Val of string Lazy.t
(*
if <> "", takes the given default value.
*)

| Doc of string
(*
if <> "", a doc string interpreted in the doc markup language.
*)

The type for what happens if the argument is absent from the cli.

type opt_kind =

| Flag
(*
without value, just a flag.
*)

| Opt
(*
with required value.
*)

| Opt_vopt of string
(*
with optional value, takes given default.
*)

The type for optional argument kinds.

type pos_kind

val pos : rev:bool -> start:int -> len:int option -> pos_kind

val pos_rev : pos_kind -> bool

val pos_start : pos_kind -> int

val pos_len : pos_kind -> int option

type t

val make : ?deprecated:string -> ?absent:string -> ?docs:string -> ?doc_envs:Env.info list -> ?docv:string -> ?doc:string -> ?env:Env.info -> string list -> t

val id : t -> int

val deprecated : t -> string option

val absent : t -> absence

val env : t -> Env.info option

val doc : t -> string

val docv : t -> string

val doc_envs : t -> Env.info list

val docs : t -> string

val opt_names : t -> string list

val opt_name_sample : t -> string

val opt_kind : t -> opt_kind

val pos_kind : t -> pos_kind

val make_req : t -> t

val make_all_opts : t -> t

val make_opt : docv:string -> absent:absence -> kind:opt_kind -> t -> t

val make_opt_all : docv:string -> absent:absence -> kind:opt_kind -> t -> t

val make_pos : docv:string -> pos:pos_kind -> t -> t

val make_pos_abs : docv:string -> absent:absence -> pos:pos_kind -> t -> t

val is_opt : t -> bool

val is_pos : t -> bool

val is_req : t -> bool

val pos_cli_order : t -> t -> int

val rev_pos_cli_order : t -> t -> int

val compare : t -> t -> int

val doclang_subst : subst:Cmdliner_manpage.subst -> t -> Cmdliner_manpage.subst

doclang_subst ~subst info adds the substitution of info to subst. Note this includes the substitutions for env if present.

val styled_deprecated : errs:Format.formatter -> subst:Cmdliner_manpage.subst -> t -> string

val styled_doc : errs:Format.formatter -> subst:Cmdliner_manpage.subst -> t -> string

type 'a conv

type e_conv =

| Conv : 'a conv -> e_conv

module Set : sig ... end