Module Cmdliner_def.Cline

Untyped command line parses.

type arg =

| O of (int * string * string option) list

| P of string list

Unconverted argument data as found on the command line.

type t

val empty : t

val add : Arg_info.t -> arg -> t -> t

val get_arg : t -> Arg_info.t -> arg

val get_opt_arg : t -> Arg_info.t -> (int * string * string option) list

val get_pos_arg : t -> Arg_info.t -> string list

val actual_args : t -> Arg_info.t -> string list

Actual command line arguments from the command line

val fold : (Arg_info.t -> arg -> 'b -> 'b) -> t -> 'b -> 'b

Deprecations

type deprecated

The type for deprecation invocations. This include both environment variable deprecations and argument deprecations.

val deprecated : env:(string -> string option) -> t -> deprecated list

deprecated ~env cli are the deprecated invocations that occur when parsing cli.

val pp_deprecated : subst:Cmdliner_manpage.subst -> deprecated Cmdliner_base.Fmt.t

pp_deprecated formats deprecations.