Module Cmdliner_info

Exit codes, environment variables, arguments, commands and eval information.

These information types gathers untyped data used to parse command lines report errors and format man pages.

module Exit : sig ... end

Exit codes.

module Env : sig ... end

Environment variables.

module Arg : sig ... end

Arguments

module Cmd : sig ... end

Commands.

module Eval : sig ... end

Evaluation.