Module Cmdliner

Declarative definition of command line interfaces.

Consult the tutorial, details about the supported command line syntax and examples of use.

Open the module to use it, it defines only three modules in your scope.

module Manpage : sig ... end

Man page specification.

module Term : sig ... end

Terms.

module Cmd : sig ... end

Commands.

module Arg : sig ... end

Terms for command line arguments.