Module Cmd.Env

Environment variable and their information.

Environment variables

type var = string

The type for environment names.

Environment variable information

type info = Term.env_info

The type for environment variable information.

val info : ?deprecated:string -> ?docs:string -> ?doc:string -> var -> info

info ~docs ~doc var describes an environment variable var such that:

doc is the man page information of the environment variable, defaults to "undocumented".

docs is the title of the man page section in which the environment variable will be listed, it defaults to Cmdliner.Manpage.s_environment.

deprecated, if specified the environment is deprecated and the string is a message output on standard error when the environment variable gets used to lookup the default value of an argument.

In doc the documentation markup language can be used with following variables:

$(env), the value of var.

The variables mentioned in info.